+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/constellation/dto/ArchiveDetailDto.java
@@ -1,7 +1,7 @@
-package com.example.starlet_be.domains.constellation.resdto;
+package com.example.starlet_be.domains.constellation.dto;
 
-import com.example.starlet_be.domains.connection.reqdto.ConnectionDto;
-import com.example.starlet_be.domains.star.resdto.StarArchiveDetailDto;
+import com.example.starlet_be.domains.connection.dto.ConnectionDto;
+import com.example.starlet_be.domains.star.dto.StarArchiveDetailDto;
 import lombok.Builder;
 import lombok.Getter;
