+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/constellation/dto/ConstellationPositionDto.java
@@ -1,4 +1,4 @@
-package com.example.starlet_be.domains.constellation.reqdto;
+package com.example.starlet_be.domains.constellation.dto;
 
 import lombok.Getter;
