+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/constellation/service/ConstellationService.java
@@ -2,23 +2,23 @@ package com.example.starlet_be.domains.constellation.service;
 
 import com.example.starlet_be.domains.connection.entity.Connection;
 import com.example.starlet_be.domains.connection.repository.ConnectionRepository;
-import com.example.starlet_be.domains.connection.reqdto.ConnectionDto;
-import com.example.starlet_be.domains.connection.resdto.StarryNightConnectionDto;
+import com.example.starlet_be.domains.connection.dto.ConnectionDto;
+import com.example.starlet_be.domains.connection.dto.StarryNightConnectionDto;
 import com.example.starlet_be.domains.constellation.entity.Constellation;
 import com.example.starlet_be.domains.constellation.repository.ConstellationRepository;
-import com.example.starlet_be.domains.constellation.reqdto.ConstellationPositionDto;
-import com.example.starlet_be.domains.constellation.reqdto.CreateConstellationDto;
-import com.example.starlet_be.domains.constellation.reqdto.UpdateConstellationInfo;
-import com.example.starlet_be.domains.constellation.resdto.ArchiveDetailDto;
-import com.example.starlet_be.domains.constellation.resdto.ArchiveDto;
-import com.example.starlet_be.domains.constellation.resdto.StarryNightConstellationDto;
+import com.example.starlet_be.domains.constellation.dto.ConstellationPositionDto;
+import com.example.starlet_be.domains.constellation.dto.CreateConstellationDto;
+import com.example.starlet_be.domains.constellation.dto.UpdateConstellationInfo;
+import com.example.starlet_be.domains.constellation.dto.ArchiveDetailDto;
+import com.example.starlet_be.domains.constellation.dto.ArchiveDto;
+import com.example.starlet_be.domains.constellation.dto.StarryNightConstellationDto;
 import com.example.starlet_be.domains.diary.entity.Color;
 import com.example.starlet_be.domains.star.entity.Star;
 import com.example.starlet_be.domains.star.repository.StarRepository;
-import com.example.starlet_be.domains.star.reqdto.StarPositionDto;
-import com.example.starlet_be.domains.star.resdto.StarArchiveDetailDto;
-import com.example.starlet_be.domains.star.resdto.StarArchiveDto;
-import com.example.starlet_be.domains.star.resdto.StarryNightStarDto;
+import com.example.starlet_be.domains.star.dto.StarPositionDto;
+import com.example.starlet_be.domains.star.dto.StarArchiveDetailDto;
+import com.example.starlet_be.domains.star.dto.StarArchiveDto;
+import com.example.starlet_be.domains.star.dto.StarryNightStarDto;
 import com.example.starlet_be.domains.user.entity.User;
 import com.example.starlet_be.domains.user.repository.UserRepository;
 import com.example.starlet_be.exception.CustomException;
