+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/email/controller/EmailController.java
@@ -1,6 +1,6 @@
 package com.example.starlet_be.domains.email.controller;
 
-import com.example.starlet_be.domains.email.reqdto.EmailAddressDto;
+import com.example.starlet_be.domains.email.dto.EmailAddressDto;
 import com.example.starlet_be.domains.email.service.EmailService;
 import com.example.starlet_be.domains.user.service.UserService;
 import com.example.starlet_be.domains.verify.service.VerifyService;
