+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/star/dto/StarArchiveDetailDto.java
@@ -1,4 +1,4 @@
-package com.example.starlet_be.domains.star.resdto;
+package com.example.starlet_be.domains.star.dto;
 
 import lombok.Builder;
 import lombok.Getter;
