+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/star/dto/StarPositionDto.java
@@ -1,4 +1,4 @@
-package com.example.starlet_be.domains.star.reqdto;
+package com.example.starlet_be.domains.star.dto;
 
 import lombok.Getter;
