+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/user/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.example.starlet_be.domains.user.controller;
 
 import com.example.starlet_be.domains.email.service.EmailService;
-import com.example.starlet_be.domains.user.reqdto.LoginDto;
-import com.example.starlet_be.domains.user.reqdto.SignUpDto;
+import com.example.starlet_be.domains.user.dto.LoginDto;
+import com.example.starlet_be.domains.user.dto.SignUpDto;
 import com.example.starlet_be.domains.user.entity.User;
 import com.example.starlet_be.exception.CustomException;
 import com.example.starlet_be.exception.ErrorCode;
