+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/user/dto/UserResDto.java
@@ -1,4 +1,4 @@
-package com.example.starlet_be.domains.user.resdto;
+package com.example.starlet_be.domains.user.dto;
 
 import lombok.Builder;
 import lombok.Data;
