+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/user/service/UserService.java
@@ -1,12 +1,11 @@
 package com.example.starlet_be.domains.user.service;
 
 import com.example.starlet_be.domains.email.entity.Email;
-import com.example.starlet_be.domains.email.repository.EmailRepository;
 import com.example.starlet_be.domains.email.service.EmailService;
-import com.example.starlet_be.domains.user.reqdto.LoginDto;
-import com.example.starlet_be.domains.user.reqdto.SignUpDto;
-import com.example.starlet_be.domains.user.resdto.LoginInfoDto;
-import com.example.starlet_be.domains.user.resdto.UserResDto;
+import com.example.starlet_be.domains.user.dto.LoginDto;
+import com.example.starlet_be.domains.user.dto.SignUpDto;
+import com.example.starlet_be.domains.user.dto.LoginInfoDto;
+import com.example.starlet_be.domains.user.dto.UserResDto;
 import com.example.starlet_be.domains.user.entity.User;
 import com.example.starlet_be.domains.user.repository.UserRepository;
 import com.example.starlet_be.domains.verify.entity.Verify;
