+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/verify/controller/VerifyApi.java
@@ -1,6 +1,6 @@
 package com.example.starlet_be.domains.verify.controller;
 
-import com.example.starlet_be.domains.verify.reqdto.PasswordResetConfirmDto;
+import com.example.starlet_be.domains.verify.dto.PasswordResetConfirmDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
@@ -9,7 +9,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 
 @Tag(name = "Verify API", description = "인증 및 보안에 관련된 API 입니다.")
 public interface VerifyApi {
